refactor: break out server benchmarks
