+++ server_benchmarks/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "server_benchmarks"
+version = "0.1.0"
+authors = ["Andrew Lamb <andrew@nerdnetworks.org>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+
+
+[dev-dependencies] # In alphabetical order
+arrow_util = { path = "../arrow_util" }
+criterion = { version = "0.3.4", features = ["async_tokio"] }
+datafusion = { path = "../datafusion" }
+flate2 = "1.0.20"
+query = { path = "../query" }
+query_tests = { path = "../query_tests" }
+server = { path = "../server" }
+test_helpers = { path = "../test_helpers" }
+tokio = { version = "1.0", features = ["macros", "time"] }
+
+
+
+[[bench]]
+name = "influxrpc"
+harness = false
